added graphics and images

+++ app/src/main/java/com/develeno/prestashopandroidclient/AboutActivity.java
@@ -11,7 +11,8 @@ public class AboutActivity extends AppCompatActivity {
     /* access modifiers changed from: protected */
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView((int) R.layout.activity_text);
+        setContentView(R.layout.activity_text);
+
     }
 
     public void back(View view) {
